% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_assignments}
\alias{format_assignments}
\title{Reformat taxonomic assignments for presentation.}
\usage{
format_assignments(
  assignments_df,
  upper_ranks = Kingdom:Phylum,
  lower_ranks = Class:Genus,
  sep = " - ",
  incomplete_label = " (unclassified)"
)
}
\arguments{
\item{assignments_df}{A data frame of taxonomic assignments.}

\item{upper_ranks}{The columns of \code{assignments_df} to use as the more
general taxon}

\item{lower_ranks}{The columns of \code{assignments_df} to use as the more
specific taxon}

\item{sep}{The separator to place between the more general and more specific
taxa}

\item{incomplete_label}{A label to append to the output if the most specific
taxon is \code{NA}. Nothing is appended if \code{incomplete_label} is set
to \code{NULL}}
}
\value{
A character vector of reformatted assignment labels.
}
\description{
Reformat taxonomic assignments for presentation.
}
\seealso{
\code{\link{split_assignments}}
}
