% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{split_assignments}
\alias{split_assignments}
\title{Split taxonomic assignments into fixed taxonomic ranks}
\usage{
split_assignments(
  assignments,
  ranks = taxonomic_ranks,
  split = "; ",
  remove = "^[kpcofgs]__"
)
}
\arguments{
\item{assignments}{Character vector of taxonomic assignments}

\item{ranks}{Character vector of taxonomic ranks, used as column names in
the result}

\item{split}{Pattern that separates the taxa in each assignment}

\item{remove}{A pattern passed to \code{stringr::str_remove}. For example,
if taxa have a prefix denoting their rank, use this pattern to remove it.
The taxa are not modified if this parameter is set to \code{NULL}}
}
\value{
A \code{tibble} containing the taxa at each rank
}
\description{
Split taxonomic assignments into fixed taxonomic ranks
}
\seealso{
\code{\link{taxonomic_ranks}}
}
